common conftest
